Fix BulkActionButtons tests
